requirements.txt
aniso8601==7.0.0
appdirs==1.4.4
asgiref==3.6.0
attrs==23.1.0
auth0-python==3.16.1
autobahn==23.1.2
Automat==22.10.0
bcrypt==4.0.1
boto3==1.17.61
botocore==1.20.112
bytecode==0.14.1
cattrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
channels==3.0.3
chardet==3.0.4
constantly==15.1.0
contextlib2==21.6.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
daphne==3.0.2
ddsketch==2.0.4
ddtrace==1.12.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
distlib==0.3.6
Django==3.2
django-cors-headers==3.7.0
django-crispy-forms==2.0
django-database-view==0.3.0
django-environ==0.4.5
django-filter==2.4.0
django-grappelli==2.15.1
django-import-export==2.5.0
django-sequences==2.6
django-split-settings==1.0.1
djangorestframework==3.12.4
djangorestframework-camel-case==1.2.0
djangorestframework-gis==0.17
drf-yasg==1.20.0
envier==0.4.0
et-xmlfile==1.1.0
fastcore==1.5.29
filelock==3.12.0
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.1.0
hyperlink==21.0.0
idna==2.8
importlib-metadata==6.0.1
incremental==22.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==0.10.0
jsonpath-ng==1.5.3
jsonschema==4.4.0
MarkupPy==1.14
MarkupSafe==2.1.2
numpy==1.24.2
odfpy==1.4.1
openpyxl==3.1.2
opentelemetry-api==1.17.0
packaging==23.1
pandas==1.2.4
Pillow==8.4.0
ply==3.11
promise==2.3
protobuf==4.22.3
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==1.7.1
pyOpenSSL==23.1.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.2
pytz==2023.3
PyYAML==6.0
requests==2.22.0
ruamel.yaml==0.17.21
Rx==1.6.3
s3transfer==0.4.2
schema==0.7.5
sentry-sdk==1.0.0
service-identity==21.1.0
simplejson==3.17.6
singledispatch==4.0.0
six==1.16.0
sqlparse==0.4.4
tablib==3.4.0
tenacity==8.2.2
text-unidecode==1.3
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.25.11
virtualenv==20.4.4
websockets==9.1
wrapt==1.15.0
xlrd==2.0.1
XlsxWriter==3.0.2
xlwt==1.3.0
xmltodict==0.13.0
zipp==3.15.0
zope.interface==6.0
